+++ src/test/kotlin/store/HoldingStoreTest.kt
@@ -17,16 +17,12 @@
 package store
 
 import impossible
-import io.gladed.watchable.Change
-import io.gladed.watchable.Watchable
-import io.gladed.watchable.WatchableValue
 import io.gladed.watchable.store.Cannot
-import io.gladed.watchable.store.Container
 import io.gladed.watchable.store.Hold
 import io.gladed.watchable.store.HoldingStore
 import io.gladed.watchable.store.Store
+import io.gladed.watchable.store.create
 import io.gladed.watchable.store.holding
-import io.gladed.watchable.toWatchableValue
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
@@ -51,15 +47,10 @@ import org.junit.Test
 import runTest
 import java.util.UUID
 
-@UseExperimental(ObsoleteCoroutinesApi::class, ExperimentalCoroutinesApi::class)
+@UseExperimental(ObsoleteCoroutinesApi::class, ExperimentalCoroutinesApi::class, FlowPreview::class)
 class HoldingStoreTest {
 
     data class Bird(val id: String = randomUuidString(), val name: String)
-    data class ContainerBird(
-        val id: String = randomUuidString(),
-        val name: WatchableValue<String>) : Container {
-        override val watchables: Watchable<Change> = name
-    }
 
     private val robin = Bird(name = "robin")
 
@@ -68,12 +59,12 @@ class HoldingStoreTest {
     private val hold = mockk<Hold>(relaxUnitFun = true)
 
     interface HoldingStoreScope : TestCoroutineScope {
-        val scopeStore: HoldingStore<Bird>
+        val holdingStore: HoldingStore<Bird>
     }
 
     private fun test(func: suspend HoldingStoreScope.() -> Unit) = runTest {
         (object : HoldingStoreScope, TestCoroutineScope by this {
-            override val scopeStore = rootStore.holding(coroutineContext) {
+            override val holdingStore = holding(rootStore) {
                 onCancel { hold.onCancel() }
                 onCreate { hold.onCreate() }
                 onRemove { hold.onRemove() }
@@ -90,33 +81,33 @@ class HoldingStoreTest {
 
     @Test fun `put an item`() = test {
         coEvery { rootStore.get(robin.id) } throws Cannot("find bird for key")
-        scopeStore.create(this).put(robin.id, robin)
+        create(holdingStore).put(robin.id, robin)
         coVerify { hold.onCreate() }
         coVerify { rootStore.put(robin.id, robin) }
     }
 
     @Test fun `release when scope completes`() = test {
         coroutineScope {
-            assertEquals(robin, scopeStore.create(this).get(robin.id))
+            assertEquals(robin, create(holdingStore).get(robin.id))
             coVerify { hold.onStart() }
         }
         coVerify { hold.onStop() }
     }
 
     @Test fun `hold only once when two scopes get`() = test {
-        assertEquals(robin, scopeStore.create(this).get(robin.id))
+        assertEquals(robin, create(holdingStore).get(robin.id))
         coVerify(exactly = 1) { hold.onStart() }
         coroutineScope {
-            assertEquals(robin, scopeStore.create(this).get(robin.id))
+            assertEquals(robin, create(holdingStore).get(robin.id))
             coVerify(exactly = 1) { hold.onStart() }
         }
     }
 
     @Test fun `release only when both scopes close`() = test {
         coroutineScope {
-            scopeStore.create(this).get(robin.id)
+            create(holdingStore).get(robin.id)
             coroutineScope {
-                scopeStore.create(this).get(robin.id)
+                create(holdingStore).get(robin.id)
             }
             coVerify(exactly = 0) { hold.onStop() }
         }
@@ -125,7 +116,7 @@ class HoldingStoreTest {
 
     @Test fun `hold+release only once when get twice from same scope`() = test {
         coroutineScope {
-            val store = scopeStore.create(this)
+            val store = create(holdingStore)
             store.get(robin.id)
             store.get(robin.id)
             coVerify(exactly = 1) { hold.onStart() }
@@ -134,8 +125,8 @@ class HoldingStoreTest {
     }
 
     @Test fun `second user must reallocate`() = test {
-        coroutineScope { scopeStore.create(this).get(robin.id) }
-        coroutineScope { scopeStore.create(this).get(robin.id) }
+        coroutineScope { create(holdingStore).get(robin.id) }
+        coroutineScope { create(holdingStore).get(robin.id) }
 
         coVerify(exactly = 2) { hold.onStart() }
         coVerify(exactly = 2) { hold.onStop() }
@@ -145,7 +136,7 @@ class HoldingStoreTest {
         coEvery { rootStore.get(eq(robin.id)) } throws Cannot("find value for that key")
 
         coroutineScope {
-            val store = scopeStore.create(this)
+            val store = create(holdingStore)
             try {
                 store.get(robin.id)
                 fail("Should have failed")
@@ -159,7 +150,7 @@ class HoldingStoreTest {
         coEvery { rootStore.get(robin.id) } throws Cannot("find value for that key")
 
         coroutineScope {
-            val store = scopeStore.create(this)
+            val store = create(holdingStore)
             try {
                 store.get(robin.id)
             } catch (c: Exception) { }
@@ -174,9 +165,9 @@ class HoldingStoreTest {
         val scope2 = CoroutineScope(coroutineContext + SupervisorJob())
 
         val birds = listOf(async {
-            scopeStore.create(scope1).get(robin.id)
+            holdingStore.create(scope1).get(robin.id)
         }, async {
-            scopeStore.create(scope2).get(robin.id)
+            holdingStore.create(scope2).get(robin.id)
         }).awaitAll()
         assertEquals(listOf(robin, robin), birds)
 
@@ -194,7 +185,7 @@ class HoldingStoreTest {
         coEvery { rootStore.get(robin.id) } coAnswers {
             mutex.withLock { robin }
         }
-        val store = scopeStore.create(this)
+        val store = create(holdingStore)
         val deferred = async {
             store.get(robin.id)
         }
@@ -207,7 +198,7 @@ class HoldingStoreTest {
 
     @Test fun `delete stops hold`() = test {
         coroutineScope {
-            val store = scopeStore.create(this)
+            val store = create(holdingStore)
             store.get(robin.id)
             store.remove(robin.id)
             coVerify { hold.onStop() }
@@ -217,7 +208,7 @@ class HoldingStoreTest {
 
     @Test fun `cannot re-put`() = test {
         coroutineScope {
-            val store = scopeStore.create(this)
+            val store = create(holdingStore)
             store.get(robin.id)
             impossible {
                 store.put(robin.id, robin.copy())
@@ -226,58 +217,15 @@ class HoldingStoreTest {
     }
 
     @Test fun `stop everybody`() = test {
-        scopeStore.create(this).get(robin.id)
-        scopeStore.stop()
+        create(holdingStore).get(robin.id)
+        holdingStore.stop()
         // Released even though scope still active
         coVerify { hold.onStop() }
     }
 
-    @UseExperimental(FlowPreview::class)
     @Test fun `keys goes to back`() = test {
         coEvery { rootStore.keys() } returns listOf(robin.id).asFlow()
-        assertEquals(listOf(robin.id), scopeStore.create(this).keys().toList())
-    }
-
-    interface ContainerWatchCreator {
-        suspend fun create(bird: ContainerBird): Hold
-    }
-
-    private val containerBird = ContainerBird(name = "one".toWatchableValue())
-    private val containerStore = mockk<Store<ContainerBird>>(relaxUnitFun = true)
-    private val containerCreator = mockk<ContainerWatchCreator>()
-
-    @Test fun `push contained changes`() = test {
-        coEvery { containerStore.get(containerBird.id) } throws Cannot("do that")
-        coEvery { containerCreator.create(containerBird) } returns hold
-        val holding = containerStore.holding(coroutineContext) { containerCreator.create(it) }
-
-        coroutineScope {
-            val scopedStore = holding.create(this)
-            scopedStore.put(containerBird.id, containerBird)
-            coVerify(exactly = 1) { containerStore.put(containerBird.id, containerBird) }
-
-            // Now change the internals of the bird which will eventually trigger re-save
-            containerBird.name.set("two")
-        }
-        // Make sure the second call happens (scope close = batch close)
-        coVerify(exactly = 2) { containerStore.put(containerBird.id, containerBird) }
-    }
-
-    @Test fun `do not push changes on deleted container item`() = test {
-        coEvery { containerStore.get(containerBird.id) } throws Cannot("do that")
-        coEvery { containerCreator.create(containerBird) } returns hold
-        val holding = containerStore.holding(coroutineContext) { containerCreator.create(it) }
-
-        coroutineScope {
-            val scopedStore = holding.create(this)
-            scopedStore.put(containerBird.id, containerBird)
-            // Change bird internals then remove it
-            containerBird.name.set("two")
-            scopedStore.remove(containerBird.id)
-            coVerify(exactly = 1) { containerStore.remove(containerBird.id) }
-        }
-        // Put only once because the bird was later removed
-        coVerify(exactly = 1) { containerStore.put(containerBird.id, containerBird) }
+        assertEquals(listOf(robin.id), create(holdingStore).keys().toList())
     }
 
     companion object {