+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Version History
 
+## 0.6.17
+* Better semantics for binding a store to a map (#95).
+* Save less when watching Container objects (#96).
+* Allow manual release of a hold with `HoldingStore.stop` (#87).
+
 ## 0.6.16
 * `holding` now relies on `HoldBuilder` for more natural composition of holds (#84).
 * Add a new way to create holding stores (#86).